requirements.txt
tensorflow-macos==2.16.2
tensorflow-hub==0.16.1
numpy==1.26.4
librosa==0.11.0
soundfile==0.13.1
scikit-learn==1.7.2
tqdm==4.67.1
h5py==3.15.1
scipy==1.15.3
numba==0.63.1
sounddevice